Add OpenGL graph mode

+++ graph_SDL.cpp
@@ -62,7 +62,7 @@ void Graph_SDL::quad(int x1,int y1,int x2,int y2,int x3,int y3,int x4,int y4,uns
     Draw_Line(one_voxel_picture,x4,y4,x1,y1,color);*/
 }
 
-void Graph_SDL::quad(int x1,int y1,int x2,int y2,int x3,int y3,int x4,int y4,unsigned short v)
+void Graph_SDL::quad(int x1,int y1,int x2,int y2,int x3,int y3,int x4,int y4,unsigned int v)
 {
     filledTrigonColor(one_voxel_picture,x1,y1,x2,y2,x3,y3,v);
     filledTrigonColor(one_voxel_picture,x3,y3,x4,y4,x1,y1,v);
